Creating fn "select-sublist".
This is useful to select a sublist of a list using a start and end index. This is nice too because if the index is greater than the size it just ignores and select until the end; the same to start. Or if start > end this fn returns an empty list.