Adding possibility to change the output prefix.
This commit keeps the current behavior where the default output prefix is " : ". It means that if you do (bot 'd/n "hello world") if you have created the bot with default params, the output in the channel would be " : hello world" like before. But now, you can set the parameter `output-prefix` to "" (or whatever you want) when you create the bot, and then the new output would be just "hello world". Or `output-prefix` + "hello world" if `output-prefix` is is different than "" (empty string).

The " : " prefix reason is just historical: it was necessary in the old legacy IRC API. Nowadays it isn't necessary at all, so is recommended you set `output-prefix` as empty string. The reason we chosen to doesn't change the default behavior is just to avoid misbehavior of old legacy codes that read/regex/interpret the channel's output.